requirements.txt
gradio
torch
torchvision
git+https://github.com/huggingface/transformers@main
sentencepiece
pillow
protobuf
accelerate